feat(resolve): implements `import-x/resolver-next`
